+++ README.md
@@ -24,11 +24,24 @@
     - [x] 유효한 회원이라면 토큰을 발행해서 로그인 상태를 유지한다
 
 3. 로그인 후 회원정보 조회 / 수정 / 삭제
-    - [ ] email 은 수정할 수 없다
-    - [ ] name / password 만 수정 가능하다
-    - [ ] 수정 api 요청을 보내는 경우는 다음 3가지다
-        - [ ] name 값 만 바꾸는 경우
-        - [ ] password 값 만 바꾸는 경우
-        - [ ] name / password 모두 바꾸는 경우
-        - [ ] 이외의 경우에는 api 를 호출하지 않는다
-    
+    - 조회
+        - [x] 로그인이 되어있지 않다면 로그인 화면으로 연결한다.
+            - [x] 나의 정보를 조회 할때, 토큰을 검사해야 한다.
+        - [x] 이메일과 이름을 표시한다.
+    - 수정(jwt 있어야 가능하다)
+        - [ ] email 은 수정할 수 없다
+        - [ ] name / password 만 수정 가능하다
+        - [ ] 수정 api 요청을 보내는 경우는 다음 3가지다
+            - [ ] name 값 만 바꾸는 경우
+            - [ ] password 값 만 바꾸는 경우
+            - [ ] name / password 모두 바꾸는 경우
+            - [ ] 이외의 경우에는 api 를 호출하지 않는다
+    - 탈퇴
+        - [ ] jwt 토큰을 복호화한 email 로 member 를 찾는다
+        - [ ] 해당 member 를 삭제하고 jwt 값을 비운다
+
+4. 로그아웃
+    - [ ] 클라이언트의 localStorage 의 jwt 값 비우고 메인 화면으로 이동
+
+5. 뷰 수정
+    - [ ] 로그인 되어있으면 로그인 -> 로그아웃 / 회원가입 노출 안함
