+++ src/main/java/wooteco/subway/service/member/MemberService.java
@@ -36,7 +36,8 @@ public class MemberService {
     }
 
     public String createToken(LoginRequest param) {
-        Member member = memberRepository.findByEmail(param.getEmail()).orElseThrow(RuntimeException::new);
+        Member member = memberRepository.findByEmail(param.getEmail())
+                .orElseThrow(() -> new RuntimeException("회원 정보가 일치하지 않습니다."));
         if (!member.checkPassword(param.getPassword())) {
             throw new RuntimeException("잘못된 패스워드");
         }
