+++ src/main/java/wooteco/subway/web/PageController.java
@@ -63,4 +63,14 @@ public class PageController {
     public String favoritesPage() {
         return "service/favorite";
     }
+
+    @GetMapping(value = "/mypage", produces = MediaType.TEXT_HTML_VALUE)
+    public String myPage() {
+        return "service/mypage";
+    }
+
+    @GetMapping(value = "/mypage-edit", produces = MediaType.TEXT_HTML_VALUE)
+    public String myPageEdit() {
+        return "service/mypage-edit";
+    }
 }
