+++ src/main/resources/static/service/js/views/MyPage.js
@@ -1,63 +1,25 @@
-import {
-  EVENT_TYPE,
-  ERROR_MESSAGE,
-  SUCCESS_MESSAGE
-} from "../../utils/constants.js";
 import api from "../../api/index.js";
-import showSnackbar from "../../lib/snackbar/index.js";
 
-function MyInfo() {
-  const $email = document.querySelector("#email");
-  const $name = document.querySelector("#name");
-  const $password = document.querySelector("#password");
-  const $signOutButton = document.querySelector("#sign-out-button");
-  const $updateButton = document.querySelector("#update-button");
+function MyPage() {
+    const $memberEmail = document.querySelector('#member-email');
+    const $memberName = document.querySelector('#member-name');
 
-  const onSignOutHandler = event => {
-    event.preventDefault();
-    if (confirm("정말 탈퇴하시겠습니까?")) {
-      api.loginMember
-        .delete()
-        .then(() => {
-          localStorage.setItem("jwt", "");
-          location.href = "/";
-        })
-        .catch(() => showSnackbar(ERROR_MESSAGE.COMMON));
+    const initMemberDetail = async () => {
+        const memberDetail = await api.loginMember.get().catch(() => {
+            alert("로그인 해주세요.");
+            return location.href = "/login";
+            }
+        );
+
+        $memberEmail.innerText = memberDetail.email;
+        $memberName.innerText = memberDetail.name;
     }
-  };
 
-  const onUpdateHandler = event => {
-    event.preventDefault();
-    const updatedInfo = {
-      name: $name.value,
-      email: $email.value,
-      password: $password.value
+    this.init = () => {
+        initMemberDetail();
     };
-    api.loginMember
-      .update(updatedInfo)
-      .then(() => {
-        showSnackbar(SUCCESS_MESSAGE.SAVE);
-      })
-      .catch(() => showSnackbar(ERROR_MESSAGE.COMMON));
-  };
-
-  const initMyInfo = () => {
-    api.loginMember
-      .get()
-      .then(member => {
-        $email.value = member.email;
-        $name.value = member.name;
-        $password.value = member.password;
-      })
-      .catch(() => showSnackbar(ERROR_MESSAGE.COMMON));
-  };
-
-  this.init = () => {
-    initMyInfo();
-    $signOutButton.addEventListener(EVENT_TYPE.CLICK, onSignOutHandler);
-    $updateButton.addEventListener(EVENT_TYPE.CLICK, onUpdateHandler);
-  };
 }
 
-const myInfo = new MyInfo();
-myInfo.init();
+const myPage = new MyPage();
+
+myPage.init();
