+++ src/main/resources/templates/service/mypage-edit.html
@@ -62,7 +62,6 @@
               class="shadow appearance-none rounded w-full py-2 px-3 text-gray-700 mb-3 leading-tight focus:outline-none focus:shadow-outline"
               id="password"
               type="password"
-              placeholder="password"
             />
           </div>
           <button
@@ -83,6 +82,6 @@
       </div>
     </div>
     <script type="module" src="/service/js/App.js"></script>
-    <script type="module" src="/service/js/views/MyPage.js"></script>
+    <script type="module" src="/service/js/views/MyPageEdit.js"></script>
   </body>
 </html>
