+++ src/main/resources/templates/service/mypage.html
@@ -14,14 +14,14 @@
         <div class="font-bold text-xl mb-4 text-center">나의 정보</div>
         <div class="mb-4">
           <div class="block text-gray-600 font-bold text-sm">email</div>
-          <div class="block text-gray-700 font-bold text-sm mb-2">eastjun@woowahan.com</div>
+          <div id="member-email" class="block text-gray-700 font-bold text-sm mb-2">eastjun@woowahan.com</div>
         </div>
         <div class="mb-4">
           <div class="block text-gray-600 font-bold text-sm">name</div>
-          <div class="block text-gray-700 font-bold text-sm mb-2">eastjun</div>
+          <div id="member-name" class="block text-gray-700 font-bold text-sm mb-2">eastjun</div>
         </div>
         <a
-          href="/mypage-edit.html"
+          href="/mypage-edit"
           class="w-full text-sm text-center bg-yellow-500 hover:bg-yellow-400 hover:text-gray-700 text-gray-800 font-semibold py-2 px-4 rounded focus:outline-none focus:shadow-outline"
           type="button"
         >
@@ -33,5 +33,6 @@
       </div>
     </div>
     <script type="module" src="/service/js/App.js"></script>
+    <script type="module" src="/service/js/views/MyPage.js"></script>
   </body>
 </html>
